Refactor jules-queue.js into focused modules
Split the monolithic `jules-queue.js` into four focused modules:
- `jules-queue-api.js`: Handles Firestore CRUD operations.
- `jules-queue-render.js`: Handles DOM rendering of queue items.
- `jules-queue-edit.js`: Manages the Edit Queue Item modal logic.
- `jules-queue-schedule.js`: Manages the Schedule Queue modal logic.

The original `jules-queue.js` is retained as a coordinator module to orchestrate these submodules and maintain backward compatibility. Updated imports in `queue-page.js`, `jules-free-input.js`, `jules-subtask-modal.js`, and `jules-modal.js` to reference the new API module where appropriate.
